Mnist_digit_Recognition/requirements.txt
streamlit
Pillow
torch
numpy
torchvision
